requirements.txt
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
fire==0.3.1
gast==0.3.3
google-auth==1.21.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.8
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mongoengine==0.20.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.3
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.2.0
